# contracts/openapi.json
{
  "components": {
    "schemas": {
      "BoundaryVersionSummary": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "effective_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "import_status": {
            "title": "Import Status",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "source_checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Checksum"
          },
          "source_crs": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Crs"
          },
          "source_file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File Name"
          },
          "source_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Format"
          },
          "source_name": {
            "title": "Source Name",
            "type": "string"
          },
          "storage_crs": {
            "title": "Storage Crs",
            "type": "string"
          },
          "validation_summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Summary"
          }
        },
        "required": [
          "id",
          "code",
          "label",
          "source_name",
          "source_file_name",
          "source_checksum",
          "source_format",
          "source_crs",
          "storage_crs",
          "import_status",
          "effective_from",
          "effective_to",
          "validation_summary"
        ],
        "title": "BoundaryVersionSummary",
        "type": "object"
      },
...
      "GeographyAliasSummary": {
        "properties": {
          "geography_unit_id": {
            "format": "uuid",
            "title": "Geography Unit Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "match_method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Method"
          },
          "match_status": {
            "title": "Match Status",
            "type": "string"
          },
          "source_code": {
            "title": "Source Code",
            "type": "string"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "source_system": {
            "title": "Source System",
            "type": "string"
          }
        },
        "required": [
          "id",
          "geography_unit_id",
          "source_system",
          "source_code",
          "source_name",
          "match_status",
          "match_method"
        ],
        "title": "GeographyAliasSummary",
        "type": "object"
      },
      "GeographyLevel": {
        "description": "Administrative hierarchy levels.\n\nAll seven levels exist in the schema from the outset. Parish and village are\ndefined but will remain empty: no parish or village boundary data has been\nsupplied, and MARS does not fabricate geography.",
        "enum": [
          "country",
          "region",
          "district",
          "county",
          "subcounty",
          "parish",
          "village"
        ],
        "title": "GeographyLevel",
        "type": "string"
      },
      "GeographyLevelCount": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "level": {
            "title": "Level",
            "type": "string"
          }
        },
        "required": [
          "level",
          "count"
        ],
        "title": "GeographyLevelCount",
        "type": "object"
      },
      "GeographyOverviewResponse": {
        "description": "Hierarchy metadata, including levels that are deliberately empty.",
        "properties": {
          "boundary_versions": {
            "items": {
              "$ref": "#/components/schemas/BoundaryVersionSummary"
            },
            "title": "Boundary Versions",
            "type": "array"
          },
          "levels": {
            "items": {
              "$ref": "#/components/schemas/GeographyLevelCount"
            },
            "title": "Levels",
            "type": "array"
          },
          "note": {
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "levels",
          "boundary_versions",
          "note"
        ],
        "title": "GeographyOverviewResponse",
        "type": "object"
      },
...
      "GeographyUnitDetail": {
        "properties": {
          "ancestors": {
            "items": {
              "$ref": "#/components/schemas/GeographyUnitSummary"
            },
            "title": "Ancestors",
            "type": "array"
          },
          "boundary_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Boundary Version Id"
          },
          "child_count": {
            "default": 0,
            "title": "Child Count",
            "type": "integer"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "effective_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective From"
          },
          "effective_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective To"
          },
          "has_geometry": {
            "default": false,
            "description": "False throughout phases 1-2; geometry is imported in Prompt 5.",
            "title": "Has Geometry",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "level": {
            "title": "Level",
            "type": "string"
          },
          "name": {
            "description": "The raw supplied name. Displayed as-is.",
            "title": "Name",
            "type": "string"
          },
          "normalised_name": {
            "description": "Lookup form. Not for display.",
            "title": "Normalised Name",
            "type": "string"
          },
          "parent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "preferred_code": {
            "title": "Preferred Code",
            "type": "string"
          },
          "unit_kind": {
            "title": "Unit Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "level",
          "unit_kind",
          "preferred_code",
          "name",
          "normalised_name",
          "parent_id",
          "depth",
          "path",
          "is_active",
          "effective_from",
          "effective_to",
          "boundary_version_id"
        ],
        "title": "GeographyUnitDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MethodDefinitionSummary": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "purpose": {
            "title": "Purpose",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/MethodVersionSummary"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "code",
          "label",
          "kind",
          "purpose"
        ],
        "title": "MethodDefinitionSummary",
        "type": "object"
      },
...
      "Page_GeographyUnitSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GeographyUnitSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Omitted when counting would be costly.",
            "title": "Total"
          }
        },
        "required": [
          "items",
          "limit",
          "offset"
        ],
        "title": "Page[GeographyUnitSummary]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "MARS bearer token": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "MARS implementation team"
    },
    "description": "MARS converts routine malaria data into explainable surveillance signals that a\nnamed person is accountable for investigating.\n\n**Scientific boundary.** Signals produced from routine e-register and HMIS data\nindicate patterns requiring investigation. They do not confirm antimalarial\nresistance. Routine data cannot distinguish recrudescence from reinfection,\nprove drug exposure or adherence, identify parasite genotype, or confirm\nmolecular markers. Externally confirmed findings - therapeutic efficacy studies\nand molecular results - are handled in a separate, separately governed lane. See\n`/api/v1/meta/evidence-lanes`.\n\n**Phase status.** This build covers phases 1 and 2: repository and runtime\nfoundation, database and governance foundation, authentication and\nauthorisation, and the organisation, facility and geography domain model. No\nencounter ingestion, episode construction, indicator, anomaly-detection, signal\nor investigation capability exists yet.\n",
    "title": "MARS - Malaria Antimalarial Resistance Surveillance",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/geography/aliases": {
      "get": {
        "description": "Return every candidate mapping for a source code.\n\nMore than one result means the mapping is ambiguous and needs review. MARS\ndoes not pick one: an ambiguous source value stays unresolved.",
        "operationId": "resolve_alias_api_v1_geography_aliases_get",
        "parameters": [
          {
            "description": "e.g. ubos_fscode, dhis2",
            "in": "query",
            "name": "source_system",
            "required": true,
            "schema": {
              "description": "e.g. ubos_fscode, dhis2",
              "title": "Source System",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source_code",
            "required": true,
            "schema": {
              "title": "Source Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeographyAliasSummary"
                  },
                  "title": "Response Resolve Alias Api V1 Geography Aliases Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "Resolve a source system's geography code",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/boundary-versions": {
      "get": {
        "operationId": "list_boundary_versions_api_v1_geography_boundary_versions_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BoundaryVersionSummary"
                  },
                  "title": "Response List Boundary Versions Api V1 Geography Boundary Versions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "Registered boundary dataset versions",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/overview": {
      "get": {
        "description": "Report unit counts per level and the boundary versions registered.\n\nLevels with no data are reported as zero rather than omitted, so an empty\nparish level reads as \"none loaded\" and not as \"level not supported\".",
        "operationId": "overview_api_v1_geography_overview_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeographyOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "Hierarchy metadata and loaded boundary versions",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/units": {
      "get": {
        "operationId": "list_units_api_v1_geography_units_get",
        "parameters": [
          {
            "description": "Filter by hierarchy level",
            "in": "query",
            "name": "level",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/GeographyLevel"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by hierarchy level",
              "title": "Level"
            }
          },
          {
            "description": "Direct children of this unit",
            "in": "query",
            "name": "parent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Direct children of this unit",
              "title": "Parent Id"
            }
          },
          {
            "in": "query",
            "name": "active_only",
            "required": false,
            "schema": {
              "default": true,
              "title": "Active Only",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_GeographyUnitSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "List geography units within your scope",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/units/{unit_id}": {
      "get": {
        "operationId": "get_unit_api_v1_geography_units__unit_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "unit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Unit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeographyUnitDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "One geography unit, with its ancestor chain",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/geography/units/{unit_id}/children": {
      "get": {
        "operationId": "get_children_api_v1_geography_units__unit_id__children_get",
        "parameters": [
          {
            "in": "path",
            "name": "unit_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Unit Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/GeographyUnitSummary"
                  },
                  "title": "Response Get Children Api V1 Geography Units  Unit Id  Children Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "Direct children of a geography unit",
        "tags": [
          "geography"
        ]
      }
    },
    "/api/v1/governance/configuration-keys": {
      "get": {
        "description": "List configuration keys.\n\nA key with no active version is returned with nulls rather than omitted: an\nungoverned parameter is a fact the programme needs to see.",
        "operationId": "list_configuration_keys_api_v1_governance_configuration_keys_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConfigurationKeySummary"
                  },
                  "title": "Response List Configuration Keys Api V1 Governance Configuration Keys Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "Governed configuration keys and which version is active",
        "tags": [
          "governance"
        ]
      }
    },
    "/api/v1/governance/methods": {
      "get": {
        "description": "List registered methods and their versions.\n\nEmpty until Prompt 13. An empty registry is the honest state: no analytical\nmethod has been defined, let alone validated or approved.",
        "operationId": "list_methods_api_v1_governance_methods_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MethodDefinitionSummary"
                  },
                  "title": "Response List Methods Api V1 Governance Methods Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "MARS bearer token": []
          }
        ],
        "summary": "The analytical method registry",
        "tags": [
          "governance"
        ]
      }
    },
...
  }
}